Moved raise_order_1D to the end of SplineObject.py
